added minor updates

+++ Jenkinsfile.groovy
@@ -50,7 +50,7 @@ pipeline {
                     def buildCommand
                     def runCommand
 
-                    String gitChanges = sh (
+                    String gitChanges = sh(
                             script: 'git diff --name-only HEAD^ HEAD',
                             returnStdout: true
                     ).trim()
@@ -59,13 +59,21 @@ pipeline {
                     for (int i = 0; i < modules.size(); i++) {
                         serviceName = modules[i].serviceName
                         if (gitChanges.contains(serviceName)) {
-                            removeContainerCommand = 'sudo docker rm \$(sudo docker stop ' + serviceName + ')'
-                            println removeContainerCommand
-                            sh removeContainerCommand
 
-                            removeImageCommand = 'sudo docker rmi medical-ledger/' + serviceName
-                            println removeImageCommand
-                            sh removeImageCommand
+                            int checkIfContainerExists = sh(
+                                    script: 'sudo docker ps -f name=' + serviceName + ' | grep -w ' + serviceName + ' | wc -l',
+                                    returnStdout: true
+                            ).trim()
+
+                            if (checkIfContainerExists != 0) {
+                                removeContainerCommand = 'sudo docker rm \$(sudo docker stop ' + serviceName + ')'
+                                println removeContainerCommand
+                                sh removeContainerCommand
+
+                                removeImageCommand = 'sudo docker rmi medical-ledger/' + serviceName
+                                println removeImageCommand
+                                sh removeImageCommand
+                            }
 
                             buildCommand = 'sudo docker build -t medical-ledger/' + serviceName + ' -f ' + modules[i].dockerPath + ' .'
                             println buildCommand
